requirements.txt
Flask
gTTS
psycopg2-binary